Add files via upload
In this program I've created rock paper scissors lizard spock! to start the program will ask your name so it can track stats, after that you are brought to th main menu where you can choose 1 or 2! 1 to start playing the game and 2 to check the stats for the games played
future work - in the future i'd like to expand upon the stats recorded in the game !